Fix header layout bug the 80+% of people with good flexbox support
This uses flexbox to fix a bug in the header layout. In Firefox, the
two columns in the header are not 50% of the width. The right-hand,
tagline container falls below, underneath the logo instead of next to
it. This only impacts people browsing using a wide screened device.

See bug report:
https://github.com/AOGPN/opengovernment.org.au/issues/32

With this commit, for people with flexbox support, flexbox overrides the hacky
inline-block layout implantation. The layout rendered is exactly the same.

Currently 80+% of people have good flexbox support, which is increasing
as people update their browsers:
http://caniuse.com/#feat=flexbox