Make the css file not minified
The master.css file is generated by a node program less. It compiles all
the .less files in the /less directory. Having this file minified into a
single line means that it's really hard to see changes in git diffs and
leads to merge conflicts. We can minify this file using a wordpress
plugin, and get sane diffs by not minifying this file.

This version of the file was generated with:

    lessc less/master.less css/master.css