Replace *Latest news* section with information about the PM’s letter to OGP
When people first arrive at the site to learn about the OGP Network,
it's important that they understand that this thing has approval from
the highest levels of Government. They need to know that the time they
invest in working towards OGP Action Plan commitments wont be wasted.

This puts the PM’s official letter (their “stamp of approval”) front and
center for people to see what they visit the home page.

Closes #105